test: fix the endpoints specifications
